geolocation-system/tests/requirements.txt
pytest==7.4.3
requests==2.31.0
Pillow==10.1.0
pytest-cov==4.1.0
